tests: reorganize command tests into subdirectory.
This makes the unit test suite much more manageable.

LP: #1651908
