+++ snapcraft/tests/commands/test_build.py
