+++ snapcraft/tests/commands/test_clean.py
