+++ snapcraft/tests/commands/test_cleanbuild.py
