+++ snapcraft/tests/commands/test_close.py
