+++ snapcraft/tests/commands/test_define.py
