+++ snapcraft/tests/commands/test_enable_ci.py
