+++ snapcraft/tests/commands/test_gated.py
