+++ snapcraft/tests/commands/test_help.py
