+++ snapcraft/tests/commands/test_history.py
