+++ snapcraft/tests/commands/test_init.py
