+++ snapcraft/tests/commands/test_list_plugins.py
