+++ snapcraft/tests/commands/test_list_registered.py
