+++ snapcraft/tests/commands/test_login.py
