+++ snapcraft/tests/commands/test_logout.py
