+++ snapcraft/tests/commands/test_pull.py
