+++ snapcraft/tests/commands/test_register.py
