+++ snapcraft/tests/commands/test_search.py
