+++ snapcraft/tests/commands/test_sign_build.py
