+++ snapcraft/tests/commands/test_snap.py
