+++ snapcraft/tests/commands/test_stage.py
