+++ snapcraft/tests/commands/test_status.py
