+++ snapcraft/tests/commands/test_update.py
