+++ snapcraft/tests/commands/test_validate.py
